Call the repair_timing_helper twice and also make the difference visible with two timing reports
That was identified to make the only real difference. An exploration was done on PcDir that checks the difference in timing from varying all the possible parameters that we can give repair_timing if we call it directly. Only calling it two times could lead to an improvement, more than two times did also not make an improvement anymore.